Optimize gol - now runs in 2 frames :) (#83)
* Enable blithog on busyloop

+++ intro/gol-cycling.c
@@ -43,7 +43,7 @@ static void ColorCyclingStep(CopInsT *ins, ColorCyclingT *rots,
     short i, n;
     ColorCyclingT *rot = &rots[j];
 
-    rot->step += (frameCount - lastFrameCount) * rot->rate;
+    rot->step += (short)(frameCount - lastFrameCount) * rot->rate;
     if (rot->step < (1 << 14))
       continue;
